requirements.txt
numpy #==1.23.3
opencv-python #==4.6.0.66
mxnet #==1.9.1
pillow #==9.2.0
tqdm #==4.64.1
Augmentor #==0.2.10
easydict #==1.9
scikit-image #==0.19.3
sklearn #==1.1.2

